+++ routes/login.py
@@ -1,10 +1,12 @@
-from database import get_db_connection
 from encrypt.encryptPass import veryfyPass
 from config.parametros import *
 
 def login(password, email, connection):
 	with connection.cursor() as cursor:
-		cursor.execute(SQL4, (email))
+		cursor.execute(SQL4, 
+				 {
+            "email": email}
+		)
 		hPass = cursor.fetchone()
 	if hPass is None:
 		return False, False